Consolidate approach to parse state, types, and sources
