+++ src/main/java/dev/ebullient/convert/tools/dnd5e/BackgroundTraits2Note.java
@@ -6,13 +6,12 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import dev.ebullient.convert.io.Tui;
-import dev.ebullient.convert.qute.QuteBase;
 import dev.ebullient.convert.qute.QuteNote;
 
 public class BackgroundTraits2Note extends Json2QuteCommon {
 
     public BackgroundTraits2Note(Tools5eIndex index) {
-        super(index, Tools5eIndexType.sourceless, null);
+        super(index, Tools5eIndexType.syntheticGroup, null);
     }
 
     public List<QuteNote> buildNotes() {
@@ -26,136 +25,139 @@ public class BackgroundTraits2Note extends Json2QuteCommon {
         return notes;
     }
 
-    @Override
-    public QuteBase build() {
-        return null;
-    }
-
     private void addIfPresent(List<QuteNote> notes, Set<String> table, String title) {
         if (table.isEmpty()) {
             return;
         }
+        boolean pushed = node == null ? parseState.push(getSources()) : parseState.push(node);
+        try {
+            List<String> text = table.stream()
+                    .map(x -> x.replaceAll("^\\|\\s*\\d+\\s*", ""))
+                    .collect(Collectors.toList());
 
-        List<String> text = table.stream()
-                .map(x -> x.replaceAll("^\\|\\s*\\d+\\s*", ""))
-                .collect(Collectors.toList());
-
-        notes.add(new QuteNote(title, null, listToTable(title, text), List.of()));
+            notes.add(new QuteNote(title, null, listToTable(title, text), List.of()));
+        } finally {
+            parseState.pop(pushed);
+        }
     }
 
     private void addIdealsIfPresent(List<QuteNote> notes) {
         if (Json2QuteBackground.ideals.isEmpty()) {
             return;
         }
+        boolean pushed = node == null ? parseState.push(getSources()) : parseState.push(node);
+        try {
+            List<String> ideals = Json2QuteBackground.ideals.stream()
+                    .map(x -> x.replace("**", ""))
+                    .map(x -> x.replaceAll("^\\|\\s*\\d+\\s*", ""))
+                    .collect(Collectors.toList());
 
-        List<String> ideals = Json2QuteBackground.ideals.stream()
-                .map(x -> x.replace("**", ""))
-                .map(x -> x.replaceAll("^\\|\\s*\\d+\\s*", ""))
-                .collect(Collectors.toList());
+            List<String> good = ideals.stream().filter(x -> x.contains("(Good)"))
+                    .collect(Collectors.toList());
+            List<String> evil = ideals.stream().filter(x -> x.contains("(Evil)"))
+                    .collect(Collectors.toList());
+            List<String> lawful = ideals.stream().filter(x -> x.contains("(Lawful)"))
+                    .collect(Collectors.toList());
+            List<String> chaotic = ideals.stream().filter(x -> x.contains("(Chaotic)"))
+                    .collect(Collectors.toList());
+            List<String> neutral = ideals.stream().filter(x -> x.contains("(Neutral)"))
+                    .collect(Collectors.toList());
+            List<String> any = ideals.stream().filter(x -> x.contains("(Any)"))
+                    .collect(Collectors.toList());
 
-        List<String> good = ideals.stream().filter(x -> x.contains("(Good)"))
-                .collect(Collectors.toList());
-        List<String> evil = ideals.stream().filter(x -> x.contains("(Evil)"))
-                .collect(Collectors.toList());
-        List<String> lawful = ideals.stream().filter(x -> x.contains("(Lawful)"))
-                .collect(Collectors.toList());
-        List<String> chaotic = ideals.stream().filter(x -> x.contains("(Chaotic)"))
-                .collect(Collectors.toList());
-        List<String> neutral = ideals.stream().filter(x -> x.contains("(Neutral)"))
-                .collect(Collectors.toList());
-        List<String> any = ideals.stream().filter(x -> x.contains("(Any)"))
-                .collect(Collectors.toList());
+            List<String> text = new ArrayList<>();
 
-        List<String> text = new ArrayList<>();
+            text.add("## Ideals");
+            text.add("");
+            text.add("| All Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^all-ideals");
+            text.add("");
+            text.add("| Chaotic Good Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^cg-ideals");
+            text.add("");
+            text.add("| Chaotic Evil Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^ce-ideals");
+            text.add("");
+            text.add("| Chaotic Neutral Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^cn-ideals");
+            text.add("");
+            text.add("| Lawful Good Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^lg-ideals");
+            text.add("");
+            text.add("| Lawful Evil Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^le-ideals");
+            text.add("");
+            text.add("| Lawful Neutral Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^ln-ideals");
+            text.add("");
+            text.add("| Neutral Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^n-ideals");
+            text.add("");
+            text.add("| Neutral Good Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^ng-ideals");
+            text.add("");
+            text.add("| Neutral Evil Ideals |");
+            text.add("|------------|");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
+            text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
+            text.add("^ne-ideals");
 
-        text.add("## Ideals");
-        text.add("");
-        text.add("| All Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^all-ideals");
-        text.add("");
-        text.add("| Chaotic Good Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^cg-ideals");
-        text.add("");
-        text.add("| Chaotic Evil Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^ce-ideals");
-        text.add("");
-        text.add("| Chaotic Neutral Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^chaotic-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^cn-ideals");
-        text.add("");
-        text.add("| Lawful Good Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^lg-ideals");
-        text.add("");
-        text.add("| Lawful Evil Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^le-ideals");
-        text.add("");
-        text.add("| Lawful Neutral Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^lawful-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^ln-ideals");
-        text.add("");
-        text.add("| Neutral Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^n-ideals");
-        text.add("");
-        text.add("| Neutral Good Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^good-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^ng-ideals");
-        text.add("");
-        text.add("| Neutral Evil Ideals |");
-        text.add("|------------|");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^neutral-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^evil-ideals)` " + " |");
-        text.add("| `dice: [](" + index.compendiumVaultRoot() + "tables/ideals.md#^universal-ideals-any)` " + " |");
-        text.add("^ne-ideals");
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Good Ideals", good));
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Evil Ideals", evil));
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Lawful Ideals", lawful));
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Chaotic Ideals", chaotic));
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Neutral Ideals", neutral));
+            maybeAddBlankLine(text);
+            text.addAll(tableSection("Universal Ideals (Any)", any));
 
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Good Ideals", good));
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Evil Ideals", evil));
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Lawful Ideals", lawful));
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Chaotic Ideals", chaotic));
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Neutral Ideals", neutral));
-        maybeAddBlankLine(text);
-        text.addAll(tableSection("Universal Ideals (Any)", any));
-
-        notes.add(new QuteNote("Ideals", null, text, List.of()));
+            notes.add(new QuteNote("Ideals", null, text, List.of()));
+        } finally {
+            parseState.pop(pushed);
+        }
     }
 
     List<String> tableSection(String title, List<String> elements) {
