+++ src/main/java/dev/ebullient/convert/tools/dnd5e/Json2QuteDeity.java
@@ -19,8 +19,7 @@ public class Json2QuteDeity extends Json2QuteCommon {
     }
 
     @Override
-    public QuteBase build() {
-
+    protected QuteBase buildQuteResource() {
         List<String> tags = new ArrayList<>(sources.getSourceTags());
 
         String pantheon = getTextOrDefault(node, "pantheon", null);
