+++ src/main/java/dev/ebullient/convert/tools/dnd5e/Tools5eIndexType.java
@@ -1,8 +1,15 @@
 package dev.ebullient.convert.tools.dnd5e;
 
-import dev.ebullient.convert.tools.IndexType;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
-public enum Tools5eIndexType implements IndexType {
+import com.fasterxml.jackson.databind.JsonNode;
+
+import dev.ebullient.convert.tools.IndexType;
+import dev.ebullient.convert.tools.NodeReader;
+
+public enum Tools5eIndexType implements IndexType, NodeReader {
     background,
     backgroundfluff,
     classtype("class"),
@@ -26,7 +33,7 @@ public enum Tools5eIndexType implements IndexType {
     optionalfeature,
     table,
     trait,
-    sourceless,
+    syntheticGroup,
     note,
     reference;
 
@@ -40,12 +47,154 @@ public enum Tools5eIndexType implements IndexType {
         this.templateName = templateName;
     }
 
+    public static final Pattern matchPattern = Pattern.compile("\\{@("
+            + Stream.of(values())
+                    .flatMap(x -> Stream.of(x.templateName, x.name()))
+                    .distinct()
+                    .collect(Collectors.joining("|"))
+            + ") ([^{}]+?)}");
+
     public String templateName() {
         return templateName;
+    }
+
+    public static Tools5eIndexType fromText(String name) {
+        return Stream.of(values())
+                .filter(x -> x.templateName.equals(name) || x.name().equalsIgnoreCase(name))
+                .findFirst().orElse(null);
     }
 
     public static Tools5eIndexType getTypeFromKey(String key) {
         String typeKey = key.substring(0, key.indexOf("|"));
         return valueOf(typeKey);
+    }
+
+    public String createKey(JsonNode x) {
+        String name = IndexElement.name.getTextOrEmpty(x);
+        String source = IndexElement.source.getTextOrEmpty(x);
+
+        switch (this) {
+            case classfeature: {
+                String classSource = IndexFields.classSource.getTextOrDefault(x, "phb");
+                return String.format("%s|%s|%s|%s|%s%s",
+                        this.name(),
+                        name,
+                        IndexFields.className.getTextOrEmpty(x),
+                        "phb".equalsIgnoreCase(classSource) ? "" : classSource,
+                        IndexFields.level.getTextOrEmpty(x),
+                        source.equalsIgnoreCase(classSource) ? "" : source)
+                        .toLowerCase();
+            }
+            case deity: {
+                return String.format("%s|%s|%s|%s",
+                        this.name(),
+                        name,
+                        IndexFields.pantheon.getTextOrEmpty(x),
+                        source)
+                        .toLowerCase();
+            }
+            case itementry: {
+                return String.format("%s|%s%s",
+                        this.name(),
+                        name,
+                        "dmg".equalsIgnoreCase(source) ? "" : "|" + source)
+                        .toLowerCase();
+            }
+            case optionalfeature: {
+                return String.format("%s|%s%s",
+                        this.name(),
+                        name,
+                        "phb".equalsIgnoreCase(source) ? "" : "|" + source)
+                        .toLowerCase();
+            }
+            case subclass: {
+                return String.format("%s|%s|%s|%s|",
+                        this.name(),
+                        name,
+                        IndexFields.className.getTextOrEmpty(x),
+                        IndexFields.classSource.getTextOrEmpty(x))
+                        .toLowerCase();
+            }
+            case subclassfeature: {
+                String classSource = IndexFields.classSource.getTextOrDefault(x, "PHB");
+                String scSource = IndexFields.subclassSource.getTextOrDefault(x, "PHB");
+                return String.format("%s|%s|%s|%s|%s|%s|%s%s",
+                        this.name(),
+                        name,
+                        IndexFields.className.getTextOrEmpty(x),
+                        "phb".equalsIgnoreCase(classSource) ? "" : classSource,
+                        IndexFields.subclassShortName.getTextOrEmpty(x),
+                        "phb".equalsIgnoreCase(scSource) ? "" : scSource,
+                        IndexFields.level.getTextOrEmpty(x),
+                        source.equalsIgnoreCase(scSource) ? "" : source)
+                        .toLowerCase();
+            }
+            case subrace: {
+                return String.format("%s|%s|%s|%s",
+                        this.name(),
+                        name,
+                        IndexFields.raceName.getTextOrEmpty(x),
+                        IndexFields.raceSource.getTextOrEmpty(x))
+                        .toLowerCase();
+            }
+            default:
+                return createKey(name, source);
+        }
+    }
+
+    public String createKey(String name, String source) {
+        if (source == null) {
+            return String.format("%s|%s", this.name(), name).toLowerCase();
+        }
+        if (this == reference) {
+            return String.format("%s|%s-%s", this.name(), name, source).toLowerCase();
+        }
+        if (this == optionalfeature) {
+            // "optionalfeature|agonizing blast",
+            // "optionalfeature|alchemical acid|uaartificer",
+            return String.format("%s|%s%s",
+                    Tools5eIndexType.optionalfeature,
+                    name,
+                    "phb".equalsIgnoreCase(source) ? "" : "|" + source)
+                    .toLowerCase();
+        }
+
+        return String.format("%s|%s|%s", this.name(), name, source).toLowerCase();
+    }
+
+    public String fromRawKey(String crossRef) {
+        return String.format("%s|%s", this.name(), crossRef).toLowerCase()
+                // NOTE: correct reference inconsistencies in the original data
+                .replaceAll("\\|phb\\|", "||")
+                .replaceAll("\\|tce\\|8\\|tce", "|tce|8");
+    }
+
+    public static String getSubclassKey(String name, String className, String classSource) {
+        return String.format("%s|%s|%s|%s|",
+                Tools5eIndexType.subclass, name, className, classSource).toLowerCase();
+    }
+
+    public static String getClassFeatureKey(String name, String featureSource, String className, String classSource,
+            String level) {
+        return String.format("%s|%s|%s|%s|%s%s",
+                Tools5eIndexType.classfeature,
+                name,
+                className,
+                "phb".equalsIgnoreCase(classSource) ? "" : classSource,
+                level,
+                featureSource.equalsIgnoreCase(classSource) ? "" : "|" + featureSource)
+                .toLowerCase();
+    }
+
+    enum IndexFields implements NodeReader {
+        level,
+        pantheon,
+        className,
+        classSource,
+        featureType,
+        raceName,
+        raceSource,
+        subclassSource,
+        subclassShortName
     }
 }
